feature-store/proposal/producer/requirements.txt
kafka-python==2.0.2
loguru==0.5.3
numpy==1.23.5
